Modify: Aside, Feed 컴포넌트화 완료 및 mock data를 이용한 여러가지 기능 구현
