+++ public/data/commentData.json
@@ -1,29 +1,20 @@
 [
   {
     "id": 1,
-    "userName": "canon_mj",
-    "img": ".././images/byounggyukim/feedimage.webp",
-    "content": "위워크에서 진행한 요리교실🍽",
-    "isLiked": "someone님외 10명이 이 게시물을 좋아합니다",
-    "comment": [
-      {
-        "id": 1,
-        "userName": "wecode",
-        "content": "Welcome to world best coding bootcamp!",
-        "isLiked": true
-      },
-      {
-        "id": 2,
-        "userName": "joonsikyang",
-        "content": "Hi there.",
-        "isLiked": false
-      },
-      {
-        "id": 3,
-        "userName": "jayPark",
-        "content": "Hey.",
-        "isLiked": false
-      }
-    ]
+    "userName": "wecode",
+    "content": "Welcome to world best coding bootcamp!",
+    "isLiked": true
+  },
+  {
+    "id": 2,
+    "userName": "joonsikyang",
+    "content": "Hi there.",
+    "isLiked": false
+  },
+  {
+    "id": 3,
+    "userName": "jayPark",
+    "content": "Hey.",
+    "isLiked": false
   }
 ]
