+++ public/data/feedData.json
@@ -1,0 +1,53 @@
+[
+  {
+    "feedPost": "wecode_bootcamp",
+    "feedImg": "images/byounggyukim/logo.jpeg",
+    "feedContent": "야호",
+    "comment": [
+      {
+        "id": 1,
+        "userName": "wecode",
+        "content": "Welcome to world best coding bootcamp!",
+        "isLiked": true
+      },
+      {
+        "id": 2,
+        "userName": "joonsikyang",
+        "content": "Hi there.",
+        "isLiked": false
+      },
+      {
+        "id": 3,
+        "userName": "jayPark",
+        "content": "Hey.",
+        "isLiked": false
+      }
+    ]
+  },
+
+  {
+    "feedPost": "wecode_bootcamp",
+    "feedImg": "images/byounggyukim/feedimage.webp",
+    "feedContent": "야호",
+    "comment": [
+      {
+        "id": 1,
+        "userName": "wecode",
+        "content": "Welcome to world best coding bootcamp!",
+        "isLiked": true
+      },
+      {
+        "id": 2,
+        "userName": "joonsikyang",
+        "content": "Hi there.",
+        "isLiked": false
+      },
+      {
+        "id": 3,
+        "userName": "jayPark",
+        "content": "Hey.",
+        "isLiked": false
+      }
+    ]
+  }
+]
