+++ src/pages/byounggyukim/Main/Aside/AsideByoungGyu.scss
@@ -1,0 +1,90 @@
+.feedRight {
+  display: flex;
+  flex-direction: column;
+  padding-top: 60px;
+  margin: 0 100px;
+  width: 20%;
+
+  .showProfile {
+    width: 100%;
+    height: 45px;
+  }
+
+  .wecode {
+    width: 35px;
+    height: 35px;
+    border-radius: 50%;
+    margin-right: 15px;
+  }
+
+  .myProfile {
+    width: 100%;
+    height: 45px;
+  }
+
+  .story {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    width: 100%;
+    height: 160px;
+    border: 1px solid #dedede;
+    background-color: white;
+    margin-top: 10px;
+    padding-left: 10px;
+    overflow: scroll;
+  }
+
+  .recommend {
+    width: 100%;
+    height: 160px;
+    border: 1px solid #dedede;
+    background-color: white;
+    margin-top: 30px;
+    padding-left: 10px;
+    overflow: scroll;
+  }
+
+  .split {
+    display: flex;
+    justify-content: space-between;
+    margin: 15px 10px;
+    margin-left: 0;
+    margin-bottom: 0;
+  }
+
+  .profile {
+    display: flex;
+    justify-content: flex-start;
+    width: 100%;
+  }
+
+  .username {
+    font-weight: bold;
+    display: flex;
+    justify-content: flex-start;
+  }
+
+  .userid {
+    font-weight: bold;
+  }
+
+  .changeColor {
+    color: gray;
+    font-weight: normal;
+  }
+
+  #seeall {
+    font-weight: bold;
+    display: flex-end;
+  }
+
+  .withoutFollow {
+    display: flex;
+    justify-content: flex-start;
+  }
+
+  .follow {
+    color: blue;
+  }
+}
