+++ src/pages/byounggyukim/Main/MainByoungGyu.scss
@@ -1,44 +1,11 @@
+.inline {
+  display: flex;
+  justify-content: center;
+  background-color: #f5f5f5;
+  width: 100vw;
+}
+
 .main {
-  .navigation {
-    display: flex;
-    justify-content: space-between;
-    padding: 10px 200px;
-    height: 50px;
-  }
-
-  .title {
-    width: 200px;
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-  }
-
-  .icon {
-    width: 30px;
-    height: 30px;
-  }
-
-  .westa {
-    font-size: 30px;
-  }
-
-  .search {
-    margin: auto 0;
-  }
-
-  .inside {
-    text-align: center;
-    height: 25px;
-    width: 250px;
-  }
-
-  .pic {
-    width: 120px;
-    display: flex;
-    justify-content: space-between;
-    margin: auto 0;
-  }
-
   .fa-solid,
   .fa-regular {
     width: 50px;
@@ -49,178 +16,101 @@
     justify-content: center;
   }
 
-  main {
-    display: flex;
-    justify-content: flex-end;
-    background-color: #f5f5f5;
-    padding-top: 60px;
-    height: 100vh;
-    width: 100vw;
-  }
-
-  .feeds {
-    margin-right: 40px;
-  }
-
-  article {
-    width: 560px;
-    border: 1px solid #dedede;
-    background-color: white;
-  }
-
-  .nicknameImage {
-    display: flex;
-    height: 40px;
-    padding: 10px 10px;
-    align-items: center;
-  }
-
-  .profileName {
-    width: 20px;
-    height: 20px;
-    border-radius: 50%;
-    margin-right: 10px;
-  }
-
   .userid {
     font-weight: bold;
     padding: auto;
   }
 
-  .feedImage {
-    width: 100%;
-    height: 60%;
-  }
-
-  .feedRight {
+  main {
     display: flex;
-    flex-direction: column;
-    margin-right: 190px;
-    width: 20%;
-  }
+    justify-content: center;
+    margin: 60px 0;
+    height: 100vh;
 
-  .likeCommentShare {
-    width: 100%;
-    display: flex;
-    justify-content: flex-start;
-  }
+    .feeds {
+      margin-right: 40px;
 
-  #heart {
-    margin-left: 10px;
-    color: red;
-  }
+      article {
+        width: 560px;
+        border: 1px solid #dedede;
+        background-color: white;
 
-  .fa-bookmark {
-    margin-left: 380px;
-  }
+        .nicknameImage {
+          display: flex;
+          height: 40px;
+          padding: 10px 10px;
+          align-items: center;
+        }
 
-  .showProfile {
-    width: 100%;
-    height: 45px;
-  }
+        .profileName {
+          width: 20px;
+          height: 20px;
+          border-radius: 50%;
+          margin-right: 10px;
+        }
 
-  .wecode {
-    width: 35px;
-    height: 35px;
-    border-radius: 50%;
-    margin-right: 15px;
-  }
+        .feedImage {
+          width: 100%;
+          height: 60%;
+        }
 
-  .myProfile {
-    width: 100%;
-    height: 45px;
-  }
+        .likeCommentShare {
+          width: 100%;
+          display: flex;
+          justify-content: flex-start;
 
-  .story {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    width: 100%;
-    height: 160px;
-    border: 1px solid #dedede;
-    background-color: white;
-    margin-top: 10px;
-    padding-left: 10px;
-    overflow: scroll;
-  }
+          #heart {
+            margin-left: 10px;
+            color: red;
+          }
 
-  .recommend {
-    width: 100%;
-    height: 160px;
-    border: 1px solid #dedede;
-    background-color: white;
-    margin-top: 30px;
-    padding-left: 10px;
-    overflow: scroll;
-  }
+          .fa-bookmark {
+            margin-left: 380px;
+          }
 
-  .split {
-    display: flex;
-    justify-content: space-between;
-    margin: 15px 10px;
-    margin-left: 0;
-    margin-bottom: 0;
-  }
+          .changeColor {
+            color: gray;
+            font-weight: normal;
+          }
+        }
+      }
 
-  .profile {
-    display: flex;
-    justify-content: flex-start;
-    width: 100%;
-  }
+      #btn {
+        width: 35px;
+        background-color: white;
+        border: none;
+        border-top: 1px solid #dedede;
+        color: #00afff;
+      }
 
-  .username {
-    font-weight: bold;
-    display: flex;
-    justify-content: flex-start;
-  }
+      #btn:hover {
+        cursor: pointer;
+      }
 
-  .changeColor {
-    color: gray;
-    font-weight: normal;
-  }
+      .commentSection {
+        padding-left: 10px;
+      }
 
-  #seeall {
-    font-weight: bold;
-    display: flex-end;
-  }
+      .addComment {
+        border: none;
+        border-top: 1px solid #dedede;
+        width: 550px;
+        padding: 0;
+        padding-left: 10px;
+        height: 50px;
+      }
 
-  .withoutFollow {
-    display: flex;
-    justify-content: flex-start;
-  }
+      .username {
+        font-weight: bold;
+        display: flex;
+        justify-content: flex-start;
+      }
 
-  #btn {
-    width: 35px;
-    background-color: white;
-    border: none;
-    border-top: 1px solid #dedede;
-    color: #00afff;
-  }
-
-  #btn:hover {
-    cursor: pointer;
-  }
-
-  .addComment {
-    border: none;
-    border-top: 1px solid #dedede;
-    width: 550px;
-    padding: 0;
-    padding-left: 10px;
-    height: 50px;
-  }
-
-  .form {
-    display: flex;
-    justify-content: space-between;
-    background-color: white;
-  }
-
-  .commentSection {
-    padding-left: 10px;
-  }
-
-  .follow {
-    color: blue;
+      .form {
+        display: flex;
+        justify-content: space-between;
+        background-color: white;
+      }
+    }
   }
 }
